Refactor and reorder
